edited

+++ hospital/migrations/20250909130834_create_medicines_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS medicines(
     name VARCHAR(255) NOT NULL,
     category medicine_category NOT NULL,
     description TEXT NOT NULL,
-    price NUMERIC(10,2) NOT NULL,
+    price FLOAT NOT NULL,
     company VARCHAR(24) NOT NULL,
     status medicine_status NOT NULL DEFAULT 'MEDICINE_STATUS_UNSPECIFIED',
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
